updated editservice function for edit
