+++ src/app/core/entity/entity-actions/entity-edit.service.ts
@@ -3,12 +3,9 @@ import { EntityMapperService } from "../entity-mapper/entity-mapper.service";
 import { Entity } from "../model/entity";
 import { EntitySchemaService } from "../schema/entity-schema.service";
 import {
-  CascadingActionResult,
   CascadingEntityAction,
 } from "./cascading-entity-action";
-import { OkButton } from "../../common-components/confirmation-dialog/confirmation-dialog/confirmation-dialog.component";
-import { KeycloakAuthService } from "../../session/auth/keycloak/keycloak-auth.service";
-import { ConfirmationDialogService } from "../../common-components/confirmation-dialog/confirmation-dialog.service";
+import { UnsavedChangesService } from "app/core/entity-details/form/unsaved-changes.service";
 
 /**
  * Safely delete an entity including handling references with related entities.
@@ -21,9 +18,33 @@ export class EntityEditService extends CascadingEntityAction {
   constructor(
     protected override entityMapper: EntityMapperService,
     protected override schemaService: EntitySchemaService,
-    private keycloakAuthService: KeycloakAuthService,
-    private confirmationDialog: ConfirmationDialogService,
+    private unsavedChanges: UnsavedChangesService,
   ) {
     super(entityMapper, schemaService);
   }
+
+// Define the generic `E extends Entity` here
+async editEntity<E extends Entity>(
+  result: any,
+  entityParam: E | E[]
+): Promise<{success: boolean, originalEntities: E[], newEntities: E[]}> {
+  if (result) {
+    let originalEntities: E[] = Array.isArray(entityParam)
+      ? entityParam
+      : [entityParam];
+    const newEntities: E[] = originalEntities.map((e) => e.copy());
+
+    for (const e of newEntities) {
+      e[result.selectedField] = result.label;
+      await this.entityMapper.save(e);
+    }
+
+    this.unsavedChanges.pending = false;
+    return {
+      success: true,
+      originalEntities,
+      newEntities
+    };
+  }
+}
 }
